requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
openai
pinecone
